+++ statham/dsl/elements/array.py
@@ -1,7 +1,8 @@
-from typing import List, TypeVar
+from typing import List, TypeVar, Union
 
-from statham.dsl.elements.base import Element
 from statham.dsl.constants import Maybe, NotPassed
+from statham.dsl.elements.base import Element
+from statham.dsl.helpers import remove_duplicates
 from statham.dsl.validation import InstanceOf
 
 
@@ -13,41 +14,53 @@ class Array(Element[List[Item]]):
 
     Requires schema element for "items" keyword as first positional
     argument. Supported validation keywords provided via keyword arguments.
-    # TODO: tuple items
-    # TODO: additionalItems
     # TODO: unqiueItems
     # TODO: contains
     """
 
-    items: Element[Item]
+    items: Union[Element[Item], List[Element]]
 
     def __init__(
         self,
-        items: Element[Item],
-        *,
-        default: Maybe[List[Item]] = NotPassed(),
         # Bad name to match JSONSchema keywords.
         # pylint: disable=invalid-name
+        items: Union[Element[Item], List[Element]],
+        *,
+        additionalItems: Union[Element, bool] = True,
+        default: Maybe[List] = NotPassed(),
         minItems: Maybe[int] = NotPassed(),
         maxItems: Maybe[int] = NotPassed(),
     ):
-        self.items = items
-        self.default = default
         # Bad name to match JSONSchema keywords.
         # pylint: disable=invalid-name
+        self.items = items
+        self.additionalItems = additionalItems
+        self.default = default
         self.minItems = minItems
         self.maxItems = maxItems
 
     @property
     def annotation(self) -> str:
-        return f"List[{self.items.annotation}]"
+        if not self.item_annotations:
+            return "List"
+        if len(self.item_annotations) == 1:
+            return f"List[{self.item_annotations[0]}]"
+        return f"List[Union[{', '.join(self.item_annotations)}]]"
+
+    @property
+    def item_annotations(self) -> List[str]:
+        """Get a list of possible type annotations."""
+        if isinstance(self.items, Element):
+            return [self.items.annotation]
+        annotations: List[str] = [item.annotation for item in self.items]
+        if self.additionalItems is True:
+            return ["Any"]
+        if isinstance(self.additionalItems, Element):
+            annotations.append(self.additionalItems.annotation)
+        if "Any" in annotations:
+            return ["Any"]
+        return remove_duplicates(annotations)
 
     @property
     def type_validator(self):
         return InstanceOf(list)
-
-    def construct(self, value, property_):
-        return [
-            self.items(item, property_.evolve(property_.name + f"[{idx}]"))
-            for idx, item in enumerate(value)
-        ]