+++ statham/dsl/validation/object.py
@@ -1,6 +1,5 @@
-from typing import Any, Optional
+from typing import Any
 
-from statham.dsl.constants import NotPassed
 from statham.dsl.exceptions import ValidationError
 from statham.dsl.validation.base import Validator
 
@@ -20,21 +19,11 @@ class AdditionalProperties(Validator):
     keywords = ("properties", "additionalProperties")
     message = "Must not contain unspecified properties. Accepts: {properties}"
 
-    @classmethod
-    def from_element(cls, element) -> Optional["AdditionalProperties"]:
-        params = list(
-            getattr(element, keyword, None) for keyword in cls.keywords
-        )
-        if None in params:
-            return None
-        params[0] = set(params[0] or {})
-        params[1] = bool(
-            True if isinstance(params[1], NotPassed) else params[1]
-        )
-        return cls(*params)
+    def error_message(self):
+        return self.message.format(properties=set(self.params["properties"]))
 
     def validate(self, value: Any):
         if self.params["additionalProperties"]:
             return
-        if set(value) - self.params["properties"]:
+        if set(value) - set(self.params["properties"]):
             raise ValidationError