+++ tests/dsl/parser/test_parse_array.py
@@ -2,8 +2,7 @@ from typing import Any, Dict
 
 import pytest
 
-from statham.dsl.elements import Array, Element, String
-from statham.dsl.exceptions import FeatureNotImplementedError
+from statham.dsl.elements import Array, Element, Integer, String
 from statham.dsl.parser import parse_element
 
 
@@ -27,19 +26,25 @@ from statham.dsl.parser import parse_element
             Array(String(), default=["foo", "bar"], minItems=1, maxItems=3),
             id="with-items-and-keywords",
         ),
+        pytest.param(
+            {
+                "type": "array",
+                "items": [{"type": "string"}, {"type": "integer"}],
+                "default": ["foo", 1],
+                "minItems": 2,
+                "additionalItems": False,
+            },
+            Array(
+                [String(), Integer()],
+                default=["foo", 1],
+                minItems=2,
+                additionalItems=False,
+            ),
+            id="with-tuple-items-and-keywords",
+        ),
     ],
 )
 def test_parse_array_produces_expected_element(
     schema: Dict[str, Any], expected: Element
 ):
     assert parse_element(schema) == expected
-
-
-def test_parse_array_tuple_items_raises():
-    with pytest.raises(FeatureNotImplementedError):
-        parse_element(
-            {
-                "type": "array",
-                "items": [{"type": "string"}, {"type": "integer"}],
-            }
-        )